Move proofs DTOs to view package
