+++ src/test/java/uk/gov/register/presentation/resource/VerifiableLogResourceTest.java
@@ -2,9 +2,9 @@ package uk.gov.register.presentation.resource;
 
 import org.junit.Test;
 import org.skife.jdbi.v2.ResultIterator;
-import uk.gov.register.presentation.EntryProof;
-import uk.gov.register.presentation.ConsistencyProof;
-import uk.gov.register.presentation.RegisterProof;
+import uk.gov.register.presentation.view.EntryProof;
+import uk.gov.register.presentation.view.ConsistencyProof;
+import uk.gov.register.presentation.view.RegisterProof;
 import uk.gov.register.presentation.dao.Entry;
 import uk.gov.register.presentation.dao.EntryQueryDAO;
 import uk.gov.verifiablelog.store.memoization.DoNothing;
